chore: publish src for nicer exploration
Sometimes a user follows a symbol to its source for further exploration
or debugging. When doing this on libraries that have source/typemap
files the IDE can use those to take the user to the original source code
rather than the built stuff that wasn't meant to be read by a human.
This only works though if said source code is available. This commit
does that.